feat: implement authentication controller and router; refactor auth routes
