+++ server/routes/authRouter.js
@@ -1,0 +1,11 @@
+// routes/authRoutes.js
+import express from 'express';
+import { register, login, googleLogin } from '../controllers/authController.js';
+
+const router = express.Router();
+
+router.post('/register', register);
+router.post('/login', login);
+router.post('/google', googleLogin);
+
+export default router;
